♻️(frontend) new component FilterOrganization
This list filter can be used for other list.
It'll be use soon on course's learners list.
